Re-generate app
