+++ package.json
@@ -1,17 +1,17 @@
 {
-  "name": "music-server",
+  "name": "music-api",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "private": true,
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "node ./bin/www"
   },
-  "keywords": [],
   "author": "Adrian Mugnolo",
   "license": "MIT",
   "dependencies": {
-    "express": "^4.13.3",
-    "lodash": "^3.10.1",
-    "morgan": "^1.6.1"
+    "debug": "~2.6.3",
+    "express": "~4.15.2",
+    "lodash": "^4.17.4",
+    "morgan": "^1.8.1"
   }
 }
